updated recipes and resources schema

+++ api/schemas/recipes.py
@@ -6,17 +6,19 @@ from .sandwiches import Sandwich
 
 
 class RecipeBase(BaseModel):
-    amount: int
+    amount: float
 
 
 class RecipeCreate(RecipeBase):
     sandwich_id: int
     resource_id: int
 
+
 class RecipeUpdate(BaseModel):
     sandwich_id: Optional[int] = None
     resource_id: Optional[int] = None
-    amount: Optional[int] = None
+    amount: Optional[float] = None
+
 
 class Recipe(RecipeBase):
     id: int
